requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[async]
asyncpg
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
alembic
